legal-agent/requirements.txt
anthropic>=0.39.0
pydantic>=2.0.0
google-genai>=1.0.0
tqdm>=4.60.0
judgeval>=0.1.0
